bd integration finished + eventviewmodel organized
