+++ MeltingApp/MeltingApp/Services/DataBaseService.cs
@@ -31,6 +31,7 @@ namespace MeltingApp.Services
             SQLiteConnection.DropTable<Token>();
             SQLiteConnection.DropTable<User>();
             SQLiteConnection.CreateTable<Event>();
+            SQLiteConnection.CreateTable<Comment>();
             SQLiteConnection.Commit();
         }
 
@@ -43,6 +44,7 @@ namespace MeltingApp.Services
             SQLiteConnection.CreateTable<User>();
             SQLiteConnection.CreateTable<Token>();
             SQLiteConnection.CreateTable<Event>();
+            SQLiteConnection.CreateTable<Comment>();
             SQLiteConnection.Commit();
         }
