+++ MeltingApp/MeltingApp/ViewModels/MainPageViewModel.cs
@@ -1,15 +1,11 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.Serialization;
-using MeltingApp.Interfaces;
 using MeltingApp.Interfaces;
 using MeltingApp.Models;
 using MeltingApp.Resources;
 using MeltingApp.Views.Pages;
 using Plugin.Media;
 using Xamarin.Forms;
-using Xamarin.Forms.PlatformConfiguration;
 using Plugin.ExternalMaps;
 
 namespace MeltingApp.ViewModels
@@ -24,13 +20,10 @@ namespace MeltingApp.ViewModels
         private string _responseMessage;
         private User _user;
         private Event _event;
-        private Event _eventSelected;
         private ImageSource _image1;
         private IEnumerable<Event> _allEvents;
         private Boolean _userAssists;
         private int _userAssistsInt;
-        private Comment _comment;
-        private IEnumerable<Comment> _allComments;
 
         public Command NavigateToCreateEventPageCommand { get; set; }
         public Command NavigateToEditProfilePageCommand { get; set; }
@@ -39,12 +32,9 @@ namespace MeltingApp.ViewModels
         public Command NavigateToStaticInfoPage { get; set; }
         public Command ShowEventCommand { get; set; }
         public Command UploadImageCommand { get; set; }
-        public Command NavigateToGetAllEventsCommand { get; set; }
-        public Command InfoEventCommand { get; set; }
+        public Command NavigateToEventViewModelCommand { get; set; }
         public Command NavigateToViewEventPageCommand { get; set; }
         public Command ConfirmAssistanceCommand { get; set; }
-        public Command CreateCommentCommand { get; set; }
-        public Command GetAllCommentsCommand { get; set; }
         public Command NavigateToFinderPage { get; set; }
         public Command OpenMapStaticFacultyCommand { get; set; }
         public Command OpenMapStaticUniversityCommand { get; set; }
@@ -58,49 +48,28 @@ namespace MeltingApp.ViewModels
             _dataBaseService = DependencyService.Get<IDataBaseService>();
             NavigateToCreateEventPageCommand = new Command(HandleNavigateToCreateEventPageCommand);
             NavigateToEditProfilePageCommand = new Command(HandleNavigateToEditProfilePageCommand);
-            NavigateToGetAllEventsCommand = new Command(HandleNavigateToGetAllEventsCommand);
+            NavigateToEventViewModelCommand = new Command(HandleNavigateToEventViewModelCommand);
             SaveEditProfileCommand = new Command(HandleSaveEditProfileCommand);
             NavigateToStaticInfoPage = new Command(HandleStaticInfoCommand);
             ViewProfileCommand = new Command(HandleViewProfileCommand);
-            InfoEventCommand = new Command(HandleInfoEventCommand);
             UploadImageCommand = new Command(HandleUploadImageCommand);
             NavigateToFinderPage = new Command(HandleFinderCommand);
-            NavigateToViewEventPageCommand = new Command(HandleNavigateToViewEventPageCommand);
             ConfirmAssistanceCommand = new Command(HandleConfirmAssistanceCommand);
             OpenMapStaticFacultyCommand = new Command(HandleOpenMapStaticFacultyCommand);
             OpenMapStaticUniversityCommand = new Command(HandleOpenMapStaticUniversityCommand);
             OpenMapEventCommand = new Command(HandleOpenMapEventCommand);
 
             Event = new Event();
-            EventSelected = new Event();
             //TODO: eliminar aquest boto
             NavigateToCreateProfilePageCommand = new Command(HandleNavigateToCreateProfilePageCommand);
             User = new User();
-            CreateCommentCommand = new Command(HandleCreateCommentCommand);
-            GetAllCommentsCommand = new Command(HandleGetAllCommentsCommand);
-            Comment = new Comment();
             FacultyStaticInfo = new StaticInfo();
             UniversityStaticInfo = new StaticInfo();
-            //Init();
+        
+           
         }
 
-        async private void Init()
-        {
-            UserAssistsInt = await _apiClientService.GetAsync<int,int>(ApiRoutes.Methods.GetUserAssistance, (isSuccess, responseMessage) =>
-            {
-                if (isSuccess)
-                {
-                    if (UserAssistsInt == 1) UserAssists = true;
-                    else UserAssists = false;
-                }
-                else
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast(responseMessage);
-                }
-            });
-        }
-
-        async void HandleConfirmAssistanceCommand()
+         async void HandleConfirmAssistanceCommand()
         {
             if (!UserAssists)
             {
@@ -163,131 +132,11 @@ namespace MeltingApp.ViewModels
                 DependencyService.Get<IOperatingSystemMethods>().ShowToast("Opening maps failed");
             }
         }
+               
 
-        async void HandleGetAllCommentsCommand()
+        void HandleNavigateToEventViewModelCommand()
         {
-            AllComments = await _apiClientService.GetAsync<IEnumerable<Comment>,IEnumerable<Comment>>(ApiRoutes.Methods.GetAllComments, (success, responseMessage) =>
-            {
-                if (success)
-                {
-                    //_navigationService.PushAsync<ViewEvent>(this);
-                }
-                else
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast(responseMessage);
-                }
-            });
-        }
-
-        async void HandleCreateCommentCommand()
-        {
-            await _apiClientService.PostAsync<Comment,Comment>(Comment, ApiRoutes.Methods.CreateComment, (isSuccess, responseMessage) =>
-            {
-                ResponseMessage = responseMessage;
-                DependencyService.Get<IOperatingSystemMethods>().ShowToast(responseMessage);
-                if (isSuccess)
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast("Comment created successfully");
-                    _navigationService.PopAsync();
-                    HandleNavigateToViewEventPageCommand();
-                }
-                else
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast(responseMessage);
-                }
-            });
-        }
-
-        void HandleInfoEventCommand()
-        {
-            Event = EventSelected;
-            //var eventscreated = _dataBaseService.GetWithChildren<Event>(e => true);
-            //Event = AllEvents.ElementAt(id);
-            _navigationService.PushAsync<ViewEvent>(this);
-        }
-
-        async void HandleNavigateToViewEventPageCommand()
-        {
-            Event = await _apiClientService.GetAsync<Event,Event>(ApiRoutes.Methods.ShowEvent, (success, responseMessage) =>
-            {
-                if (success)
-                {
-                    HandleGetAllCommentsCommand();
-                    _navigationService.PushAsync<ViewEvent>(this);
-                }
-                else
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast(responseMessage);
-                }
-            });
-        }
-
-        async void HandleNavigateToGetAllEventsCommand()
-        {
-            AllEvents = await _apiClientService.GetAsync<IEnumerable<Event>, IEnumerable<Event>>(ApiRoutes.Methods.GetAllEvents, (success, responseMessage) =>
-            {
-                if (success)
-                {
-                    _navigationService.PushAsync<EventList>(this);
-                }
-                else
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast(responseMessage);
-                }
-            });
-
-            saveEventsInDB(AllEvents); //guardem tots els events a la base de dades
-
-        }
-
-        /// <summary>
-        /// guardem tots els events a la base de dades, i les seves referencies als seus usuaris creadors, si
-        /// el creador no es troba a la base de dades, l'obtenim, l'afegim a la taula users i li posem la referencia a l'event
-        /// </summary>
-        /// <param name="AllEvents"></param>
-        void saveEventsInDB(IEnumerable<Event>AllEvents)
-        {
-            var allevents_before = _dataBaseService.GetCollectionWithChildren<Event>(e => true);
-            for (int i = 0; i < AllEvents.Count(); i++)
-            {
-                //comprovar si el event ja esta a la bd
-                var eventt = AllEvents.ElementAt(i);
-                bool b = false;
-                for (int j = 0; j < allevents_before.Count() && !b ; j++)
-                {
-                    if (allevents_before.ElementAt(j).id == eventt.id)
-                    {
-                        b = true;
-                    }
-                }
-                if (!b)
-                {
-                    //si levent no esta a la bd
-                    var eventToSave = AllEvents.ElementAt(i);
-                    _dataBaseService.UpdateWithChildren<Event>(eventToSave);
-                }
-                
-            }
-            var allevents_after = _dataBaseService.GetCollectionWithChildren<Event>(e => true);
-        }
-        
-        async void getAndSaveProfileInDB(int iduser)
-        {
-            bool b = false;
-            var meltingUriParser = new MeltingUriParser();
-            meltingUriParser.AddParseRule(ApiRoutes.UriParameters.UserId, $"{iduser}");
-
-            User = await _apiClientService.GetAsync<User, User>(ApiRoutes.Methods.GetProfileUser, (success, responseMessage) =>
-            {
-                if (success)
-                {
-                    b = true;
-                }
-            }, meltingUriParser);
-            if (b)
-            {
-                _dataBaseService.UpdateWithChildren<User>(User);
-            }
+            EventViewModel evm = new EventViewModel();
         }
 
         private void HandleNavigateToCreateProfilePageCommand()
@@ -412,36 +261,6 @@ namespace MeltingApp.ViewModels
             {
                 _event = value;
                 OnPropertyChanged(nameof(Event));
-            }
-        }
-
-        public Event EventSelected
-        {
-            get { return _eventSelected; }
-            set
-            {
-                _eventSelected = value;
-                OnPropertyChanged(nameof(EventSelected));
-            }
-        }
-
-        public Comment Comment
-        {
-            get { return _comment; }
-            set
-            {
-                _comment = value;
-                OnPropertyChanged(nameof(Comment));
-            }
-        }
-
-        public IEnumerable<Comment> AllComments
-        {
-            get { return _allComments; }
-            set
-            {
-                _allComments = value;
-                OnPropertyChanged(nameof(AllComments));
             }
         }
