+++ MeltingApp/MeltingApp/Views/Pages/EventList.xaml.cs
@@ -16,7 +16,7 @@ namespace MeltingApp.Views.Pages
 		{
 			InitializeComponent ();
 		    NavigationPage.SetHasNavigationBar(this, false);
-		    BindingContext = new MainPageViewModel();
+		  
         }
 	}
 }
