+++ MeltingApp/MeltingApp/Views/Pages/MainPage.xaml
@@ -12,7 +12,7 @@
             <Button Text="Static Info" Command="{Binding NavigateToStaticInfoPage}"></Button>
             <Button Text="Finder" Command="{Binding NavigateToFinderPage}"></Button>
             <Button Text="Create Event" Command="{Binding NavigateToCreateEventPageCommand}"></Button>
-            <Button Text="Get All Events" Command="{Binding NavigateToGetAllEventsCommand}"></Button>
+            <Button Text="Get All Events" Command="{Binding NavigateToEventViewModelCommand}"></Button>
             <Button Text="View Event" Command="{Binding NavigateToViewEventPageCommand}"></Button>
             <Button x:Name="ViewButton" Text="View Profile" Command="{Binding ViewProfileCommand}"></Button>
         </StackLayout>
